+++ reservation-service/src/main/webapp/views/detail.jsp
@@ -24,7 +24,7 @@
                         <a href="/" class="lnk_logo" title="네이버"> <span class="spr_bi ico_n_logo">네이버</span> </a>
                         <a href="/" class="lnk_logo" title="예약"> <span class="spr_bi ico_bk_logo">예약</span> </a>
                     </h1>
-                    <a class="btn_my"> <span title="내 예약">MY</span> </a>
+                    <a class="btn_my" href = "/my"> <span title="내 예약">MY</span> </a>
                 </header>
                 <div class="pagination">
                     <div class="bg_pagination"></div>
